refactor: modify Evaluator part for sunset petsard metadata and util
